+++ R/importance_sampling.R
@@ -19,7 +19,7 @@ importance_sampling <- function(log_ratios, method, ...) {
 importance_sampling.array <-
   function(log_ratios, method,
            ...,
-           r_eff = NULL,
+           r_eff = 1,
            cores = getOption("mc.cores", 1)) {
     cores <- loo_cores(cores)
     stopifnot(length(dim(log_ratios)) == 3)
@@ -36,7 +36,7 @@ importance_sampling.array <-
 importance_sampling.matrix <-
   function(log_ratios, method,
            ...,
-           r_eff = NULL,
+           r_eff = 1,
            cores = getOption("mc.cores", 1)) {
     cores <- loo_cores(cores)
     assert_importance_sampling_method_is_implemented(method)
@@ -49,7 +49,7 @@ importance_sampling.matrix <-
 #' @inheritParams psis
 #' @export
 importance_sampling.default <-
-  function(log_ratios, method, ..., r_eff = NULL) {
+  function(log_ratios, method, ..., r_eff = 1) {
     stopifnot(is.null(dim(log_ratios)) || length(dim(log_ratios)) == 1)
     assert_importance_sampling_method_is_implemented(method)
     dim(log_ratios) <- c(length(log_ratios), 1)
@@ -128,7 +128,7 @@ implemented_is_methods <- function() c("psis", "tis", "sis")
 #'   but unnormalized.
 #' @param pareto_k Vector of GPD k estimates.
 #' @param tail_len Vector of tail lengths used to fit GPD.
-#' @param r_eff Vector of relative MCMC n_eff for `exp(log lik)`
+#' @param r_eff Vector of relative MCMC ESS (n_eff) for `exp(log lik)`
 #' @template is_method
 #' @return A list of class `"psis"` with structure described in the main doc at
 #'   the top of this file.
@@ -153,7 +153,7 @@ importance_sampling_object <-
     out <- structure(
       list(
         log_weights = unnormalized_log_weights,
-        diagnostics = list(pareto_k = pareto_k, n_eff = NULL)
+        diagnostics = list(pareto_k = pareto_k, n_eff = NULL, r_eff = r_eff)
       ),
       # attributes
       norm_const_log = norm_const_log,
@@ -184,6 +184,7 @@ do_importance_sampling <- function(log_ratios, r_eff, cores, method) {
   assert_importance_sampling_method_is_implemented(method)
   N <- ncol(log_ratios)
   S <- nrow(log_ratios)
+  k_threshold <- ps_khat_threshold(S)
   tail_len <- n_pareto(r_eff, S)
 
   if (method == "psis") {
@@ -223,7 +224,7 @@ do_importance_sampling <- function(log_ratios, r_eff, cores, method) {
 
   log_weights <- psis_apply(lw_list, "log_weights", fun_val = numeric(S))
   pareto_k <- psis_apply(lw_list, "pareto_k")
-  throw_pareto_warnings(pareto_k)
+  throw_pareto_warnings(pareto_k, k_threshold)
 
   importance_sampling_object(
     unnormalized_log_weights = log_weights,