+++ man-roxygen/loo-and-psis-references.R
@@ -5,7 +5,7 @@
 #' ([journal version](https://link.springer.com/article/10.1007/s11222-016-9696-4),
 #'  [preprint arXiv:1507.04544](https://arxiv.org/abs/1507.04544)).
 #'
-#' Vehtari, A., Simpson, D., Gelman, A., Yao, Y., and Gabry, J. (2019).
+#' Vehtari, A., Simpson, D., Gelman, A., Yao, Y., and Gabry, J. (2022).
 #' Pareto smoothed importance sampling.
 #' [preprint arXiv:1507.02646](https://arxiv.org/abs/1507.02646)
-#'
+#'